Date 23/09 (Done)

+++ src/main/java/Date_1809/utils/BMI.java
@@ -21,6 +21,6 @@ public class BMI {
             throw new InvalidHeightException();
         }
 
-        return w / (h * h);
+        return (double) w / (h * h);
     }
 }
